+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## Unreleased
+## Added
+* replaced `getAuthenticationSpecification` with `authenticationSpecification`
+
 ## [1.1.1] - 2018-05-31
 ## Fixed
 * Include peer-dependencies koop and and koop-output-geoservices